TECHDEBT: Reorganized My Pokemons and Pokemon Inspector folders - Removed unused folders
