--- src/components/MyPokemons/MyPokemons.js
@@ -1,12 +1,0 @@
-import PokemonsList from "../PokemonsList/PokemonsList";
-import styles from "./MyPokemons.module.scss";
-
-const MyPokemons = () => {
-  return (
-    <div className={styles.myPokemonsPageContainer}>
-      <PokemonsList />
-    </div>
-  );
-};
-
-export default MyPokemons;
