+++ src/components/Pokemons/MyPokemons/AddPokemon/AddPokemon.module.scss
@@ -1,4 +1,4 @@
-@import "../../css/colors.scss";
+@import "../../../../css/colors.scss";
 
 .addPokemonForm {
   margin: 0 auto;
