+++ src/components/Pokemons/MyPokemons/CustomPokemon/CustomPokemon.js
