+++ src/components/Pokemons/MyPokemons/CustomPokemon/CustomPokemon.module.scss
