+++ src/components/Pokemons/MyPokemons/MyPokemonsList/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MyPokemonsList";
