+++ src/components/Pokemons/PokemonInspector/Pokemon/Pokemon.js
