+++ src/components/Pokemons/PokemonInspector/Pokemon/Pokemon.module.scss
