+++ src/components/Pokemons/PokemonInspector/Pokemon/index.js
