+++ src/components/Pokemons/PokemonInspector/PokemonInspector/PokemonInspector.module.scss
@@ -1,4 +1,4 @@
-@import "../../css/colors.scss";
+@import "../../../../css/colors.scss";
 
 .buyPokemonsContainer {
   display: flex;
