+++ src/components/Pokemons/PokemonInspector/PokemonInspector/index.js
@@ -1,0 +1,1 @@
+export { default } from "./PokemonInspector";
