--- src/components/PokemonsList/PokemonsList.js
@@ -1,7 +1,0 @@
-import styles from "./PokemonsList.module.scss";
-
-const PokemonsList = (props) => {
-  return <div className={styles.pokemonsListContainer}></div>;
-};
-
-export default PokemonsList;
