--- src/components/PokemonsList/PokemonsList.module.scss
@@ -1,12 +1,0 @@
-.pokemonListContainer {
-  width: 60rem;
-  align-self: center;
-  height: 10rem;
-  background-color: green;
-  display: flex;
-}
-
-.loading {
-  font-weight: 600;
-  font-size: 1rem;
-}
